streaming: fix sample conversion hang at 65536 or more samples
slowPath_convert used a uint16_t loop counter against a uint32_t
count, so at 65536 samples the counter wrapped before reaching the
bound and the loop never ended. The neighbouring unzip and zip paths
already use uint32_t. Adds the unit test that would have caught it.

Fixes #195. This is the root cause behind the symptom in #125.

+++ src/streaming/samplesConversion.h
@@ -91,7 +91,7 @@ template<uint32_t srcCount, class DestT, class SrcT> static void fastPath_conver
 // dynamic iteration count, will generate extra SIMD instructions for various last iteration counts
 template<class DestT, class SrcT> static void slowPath_convert(DestT* dest, const SrcT* src, uint32_t srcCount)
 {
-    for (uint16_t i = 0; i < srcCount; ++i)
+    for (uint32_t i = 0; i < srcCount; ++i)
         Rescale(dest[i], src[i]);
 }
